refactor: now kebab-case must be enforced
