+++ src/integrations/db/db.integration.ts
@@ -1,8 +1,8 @@
 import { Integration } from "../integration.interface";
 import { DbClient } from "./db.client";
 import { DbAction } from './db.types';
-import { addGuildRepository, removeGuildRepository, getGuildRepositories } from "./guildRepositories";
-import { getUserMapping, setUserMapping, deleteUserMapping } from "./userMappings";
+import { addGuildRepository, removeGuildRepository, getGuildRepositories } from "./guild-repositories";
+import { getUserMapping, setUserMapping, deleteUserMapping } from "./user-mappings";
 
 export class DBIntegration implements Integration {
   name = 'db';
@@ -25,17 +25,17 @@ export class DBIntegration implements Integration {
   async execute(action: DbAction, payload: any) {
     switch (action) {
       case 'add_guild_repository':
-        return addGuildRepository(this.getClient().getDbClient(), payload);
+        return addGuildRepository(this.getClient().getDb(), payload);
       case 'remove_guild_repository':
-        return removeGuildRepository(this.getClient().getDbClient(), payload);
+        return removeGuildRepository(this.getClient().getDb(), payload);
       case 'get_guild_repositories':
-        return getGuildRepositories(this.getClient().getDbClient(), payload);
+        return getGuildRepositories(this.getClient().getDb(), payload);
       case 'get_user_mapping':
-        return getUserMapping(this.getClient().getDbClient(), payload);
+        return getUserMapping(this.getClient().getDb(), payload);
       case 'set_user_mapping':
-        return setUserMapping(this.getClient().getDbClient(), payload);
+        return setUserMapping(this.getClient().getDb(), payload);
       case 'delete_user_mapping':
-        return deleteUserMapping(this.getClient().getDbClient(), payload);
+        return deleteUserMapping(this.getClient().getDb(), payload);
       default:
         throw new Error(`Action ${action} not supported`);
     }
