--- src/integrations/db/guildRepositories/index.ts
@@ -1,5 +1,0 @@
-export * from "./actions/addGuildRepository.action";
-export * from "./actions/removeGuildRepository.action";
-export * from "./actions/getGuildRepositories.action";
-
-export * from "./guildRepositories.types";
