+++ src/integrations/db/index.ts
@@ -1,6 +1,6 @@
 export * from './db.client';
 export * from './db.types';
 export * from './db.integration';
-export * from './guildRepositories';
-export * from './userMappings';
+export * from './guild-repositories';
+export * from './user-mappings';
