+++ src/integrations/db/user-mappings/actions/delete-user-mapping.action.ts
@@ -1,8 +1,8 @@
 import type { DbClient } from '../../db.client';
-import type { DeleteUserMappingPayload } from '../userMappings.types';
+import type { DeleteUserMappingPayload } from '../user-mappings.types';
 
 export function deleteUserMapping(
-  db: ReturnType<DbClient['getDbClient']>,
+  db: ReturnType<DbClient['getDb']>,
   payload: DeleteUserMappingPayload,
 ): void {
   db.prepare(`DELETE FROM user_mappings WHERE discord_id = ?`).run(
