+++ src/integrations/db/user-mappings/index.ts
@@ -1,0 +1,5 @@
+export * from "./actions/delete-user-mapping.action";
+export * from "./actions/get-user-mapping.action";
+export * from "./actions/set-user-mapping.action";
+
+export * from "./user-mappings.types";
