--- src/integrations/db/userMappings/index.ts
@@ -1,5 +1,0 @@
-export * from "./actions/deleteUserMapping.action";
-export * from "./actions/getUserMapping.action";
-export * from "./actions/setUserMapping.action";
-
-export * from "./userMappings.types";
