+++ src/integrations/github/issue/index.ts
@@ -1,5 +1,5 @@
-export * from "./actions/assignIssue.action";
-export * from "./actions/createIssue.action";
-export * from "./actions/getIssues.action";
+export * from "./actions/assign-issue.action";
+export * from "./actions/create-issue.action";
+export * from "./actions/get-issues.action";
 
 export * from "./issue.types";
