+++ src/interfaces/bot/commands/index.ts
@@ -8,11 +8,11 @@ import {
 import * as helps from './helps.command';
 import * as ping from './ping.command';
 import * as status from './status.command';
-import * as createIssue from './createIssue.command';
-import * as addRepository from './addRepository.command';
-import * as assignIssue from './assignIssue.command';
-import * as linkGithub from './linkGithub.command';
-import * as unlinkGithub from './unlinkGithub.command';
+import * as createIssue from './create-issue.command';
+import * as addRepository from './add-repository.command';
+import * as assignIssue from './assign-issue.command';
+import * as linkGithub from './link-github.command';
+import * as unlinkGithub from './unlink-github.command';
 
 export interface Command {
   data: SlashCommandBuilder | SlashCommandOptionsOnlyBuilder;
