+++ src/interfaces/bot/handlers/client-ready.handler.ts
