+++ src/interfaces/bot/handlers/index.ts
@@ -1,8 +1,8 @@
 import { Client } from 'discord.js';
-import { handleClientReady } from './clientReady.handler';
-import { handleInteraction } from './interactionCreate.handler';
-import { handleGuildCreate } from './guildCreate.handler';
-import { handleMessageCreate } from './messageCreate.handler';
+import { handleClientReady } from './client-ready.handler';
+import { handleInteraction } from './interaction-create.handler';
+import { handleGuildCreate } from './guild-create.handler';
+import { handleMessageCreate } from './message-create.handler';
 
 export function registerHandlers(client: Client): void {
   client.on('clientReady', handleClientReady);
